% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bolivia_new.R
\docType{data}
\name{Bolivia_new}
\alias{Bolivia_new}
\title{Altiplane Peru Bolivia Rainfall Information}
\format{A list rain and cov:
\itemize{
  \item Bolivia_new$rain: 72 rainfall station is a sp object.
  \item Bolivia_new$x: longitude raster.
  \item Bolivia_new$y: latitude raster.
  \item Bolivia_new$x: rainfall raster.
  \item Bolivia_new$dem: Digital Elevation Model raster.}}
\usage{
Bolivia_new
}
\description{
A dataset containing rainfall information and other spatial covariables.
Dorado. The variables are as follows:
}
\examples{
library(Dorado)
data("Bolivia_new")
}
\keyword{datasets}

