% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dorado.R
\docType{data}
\name{Dorado}
\alias{Dorado}
\title{Peruvian Rainfall Information}
\format{A dataset containing rainfall information and TRMM2A25:
\itemize{
  \item Bolivia_new$rain: 83 rainfall station is a sp object.
  \item Bolivia_new$TRMM: raster precipitation.
   }}
\usage{
Dorado
}
\description{
Peruvian Rainfall Information
}
\examples{
library(Dorado)
data("Dorado")
}
\keyword{datasets}
