% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_CHIRP.R
\name{downloadCHIRPS}
\alias{downloadCHIRPS}
\title{Download CHIRPS}
\usage{
downloadCHIRPS(res = "p05", day = as.Date("2015-10-01"),
  saveCHIRPS = "~/", BBlonMin = -86, BBlonMax = -66, BBlatMin = -19.25,
  BBlatMax = 1.25)
}
\arguments{
\item{res}{CHIRP type 005 or 025.}

\item{day}{Dia que se desea descargar}

\item{saveCHIRPS}{Lugar en donde se guardara el archivo}

\item{BBlonMin}{Crop-Box long min}

\item{BBlonMax}{Crop-Box long max}

\item{BBlatMin}{Crop-Box lat min}

\item{BBlatMax}{Crop-Box lat max}
}
\description{
Download CHIRPS
}
\author{
Cesar  Aybar
}
