% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean_R.R
\name{Euclidean_R}
\alias{Euclidean_R}
\title{Estimate Euclidean Distance for sp Objects}
\usage{
Euclidean_R(sp, raster)
}
\arguments{
\item{sp}{Is a object sp (\code{SpatialPolygonsDataFrame}, \code{SpatialLineDataFrame}, \code{SpatialPointsDataFrame}).}

\item{raster}{Is a RasterLayer object.}
}
\description{
This function implement a quick method for estimate the euclidean distance using
a RasterLayer object for defined the boundaries.
}
\details{
This function use the raster only for extract the coordinates referent system and geometries attributes,
the values are not used.
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}

