% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REuclidean.R
\name{EuclideanRaster}
\alias{EuclideanRaster}
\title{Estimación de la distancia Euclideana en Raster}
\usage{
EuclideanRaster(shp, rasterModel, output)
}
\arguments{
\item{shp:}{Es un objeto de clase SpatialPolygonsDataFrame, SpatialLineDataFrame, SpatialPointsDataFrame.}

\item{rasterModel:}{Es un objeto clase Raster}

\item{Es}{el carpeto de destino}
}
\description{
Estimación de la distancia Euclideana en Raster
}
\author{
Cesar Luis Aybar Camacho
}

