% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fast_Extract.R
\name{Extract_Fast_Multiple}
\alias{Extract_Fast_Multiple}
\title{Fast Extract RasterLayer data through shapefile with many geometries}
\usage{
Extract_Fast_Multiple(R1, shpR, dirtrash, fun = list(mean, max, min))
}
\description{
Fast Extract RasterLayer data through shapefile with many geometries
}

