% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fast_Extract.R
\name{Extract_Fast_One}
\alias{Extract_Fast_One}
\title{Fast Extract RasterLayer data through shapefile with one geometry}
\usage{
Extract_Fast_One(R1, shpR, Routput, fun = mean)
}
\description{
Fast Extract RasterLayer data through shapefile with one geometry
}

