% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitVariogram.R
\name{FitVariogram}
\alias{FitVariogram}
\title{Fit Varigoram using autofitVariogram modified}
\usage{
FitVariogram(formula, input_data, model = c("Sph", "Exp", "Gau", "Ste"),
  kappa = c(0.05, seq(0.2, 2, 0.1), 5, 10), fix.values = c(NA, NA, NA),
  verbose = FALSE, GLS.model = NA, start_vals = c(NA, NA, NA),
  miscFitOptions = list(), boundaries, ...)
}
\arguments{
\item{formula}{that defines the dependent variable as a linear model
of independent variables; suppose the dependent variable has
name 'z', for ordinary and simple kriging use the formula
'z~1'; for simple kriging also define 'beta' (see below); for
universal kriging, suppose 'z' is linearly dependent on 'x'
and 'y', use the formula 'z~x+y'.}

\item{input_data}{An object of \link[sp]{SpatialPointsDataFrame-class}.}

\item{model}{The list of variogrammodels that will be tested.}

\item{kappa}{Smoothing parameter of the Matern model. Provide a list if you want to check
more than one value.}

\item{fix.values}{Can be used to fix a variogram parameter to a certain value. It
consists of a list with a length of three. The items describe the
fixed value for the nugget, range and sill respectively. They need to be given in that order.
Setting the value to NA means that the value is not fixed.}

\item{verbose}{logical, if TRUE the function will give extra feedback on the fitting process}

\item{GLS.model}{If a variogram model is passed on through this parameter a Generalized Least Squares
sample variogram is calculated.}

\item{start_vals}{Can be used to give the starting values for the variogram fitting. The items describe the
fixed value for the nugget, range and sill respectively. They need to be given in that order.
Setting the value to NA means that the value will be automatically chosen.}

\item{miscFitOptions}{A list with named arguments that provide additional control over the fitting process.
For example: \code{list(merge.small.bins = TRUE)}. If the list is empty, autofitVariogram
uses default values. The following parameters can be set:
\describe{
 \item{\code{merge.small.bins}:}{logical, when TRUE, the function checks if there are bins with less than 5 points.
 If so, the first two bins are merged and the check is repeated. This is done until all bins have more
 than \code{min.np.bin} points.}
 \item{\code{min.np.bin}:}{integer, the minimum number of points allowed in a bin before
 we start merging bins. See also \code{merge.small.bins}.}
 }}

\item{boundaries}{Are the Boundaries for the bins in km.}

\item{\code{...}}{parameters that are passed on to \link[gstat]{variogram} variogram when calculating the sample variogram}
}
\value{
An object of type \code{autofitVariogram} is returned. This object contains the experimental variogram,
 the fitted variogram model and the sums of squares (\code{sserr}) between the sample variogram and the
 fitted variogram model.
}
\description{
This function is a very slight variation about the function autofitVariogram of packages
automap. Only added the param boundaries so the user can define it.
}
\details{
Geostatistical routines are used from package \code{gstat}.
 A few simple choices are made when estimating the inital guess for \code{fit.variogram}.
 The initial sill is estimated as the \code{mean} of the \code{max} and the \code{median}
 of the semi-variance. The inital range is defined as 0.10 times the diagonal of the bounding
 box of the data. The initial nugget is defined as the \code{min} of the the semi-variance.
 There are five different types of models that are often used:
 \describe{
   \item{Sph}{A shperical model.}
   \item{Exp}{An exponential model.}
   \item{Gau}{A gaussian model.}
   \item{Mat}{A model of the Matern familiy}
   \item{Ste}{Matern, M. Stein's parameterization}
   }
A list of all permitted variogram models is available by typing vgm() into the R console.
     \code{autofitVariogram} iterates over the variogram models listed in \code{model} and picks the model
     that has the smallest residual sum of squares with the sample variogram. For the Matern model, all the
     kappa values in \code{kappa} are tested.

     Note that when using the power model, and not specifying starting values yourself, the sill is set to 1,
     the range to 1 and the nugget to 0. This is because the normal initial values for those paramters don't
     work well with the power model. I consider this a temporary solution, any suggestions are appreciated.

It is possible to pass anisotropy parameters to \code{autofitVariogram}. However, \code{autofitVariogram} does not fit anisotropic variogram models. The function sees the anisotropic sample variogram as one big sample variogram. So it fits an average isotropic variogram model from the anisotropic sample variogram. A warning is issued when a users passes \code{alpha} to \code{autofitVariogram}.
}
\author{
Cesar Aybar
}
\seealso{
\code{\link[gstat]{fit.variogram}}, \code{\link{autoKrige}}, \code{\link{posPredictionInterval}}
}

