% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateXYraster.R
\name{generateXYraster}
\alias{generateXYraster}
\title{Estimación de raster de Longitud y latitud}
\usage{
generateXYraster(x)
}
\arguments{
\item{x:}{Raster object}
}
\description{
Estimación de raster de Longitud y latitud
}
\examples{
data(Prec_Bolivia)
generateXYraster(Prec_Bolivia)
}
\author{
Cesar Luis Aybar Camacho
}

