% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_me_the_data.R
\name{give_me_the_data}
\alias{give_me_the_data}
\title{Extract Dorado data}
\usage{
give_me_the_data(data = "Titicaca", output = "~/")
}
\arguments{
\item{data}{Data to extract}

\item{output}{is the folder directory to save the data.}
}
\description{
Extract Dorado data
}
\examples{
give_me_the_data(Titicaca)
give_me_the_data(Dorado)
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}

