% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDW.R
\name{IDW}
\alias{IDW}
\title{IDW Optimizing inverse distance weighting power}
\usage{
IDW(gauge, cov, formula, idpR = seq(0.8, 3.5, 0.1))
}
\arguments{
\item{gauge}{Is an object of SpatialPointsDataFrame class.}

\item{cov}{Is An object of RasterLayer.}

\item{formula}{that defines the dependent variable as a linear model
of independent variables; suppose the dependent variable has
name 'z', for Regression Inverse Distance Weigthing (RIDW) use the formula
'z~x+y+....', you do not need define.}

\item{idpR}{Is vector numeric of the power coeficient to evaluate.}
}
\value{
a List that contains: \code{Interpol} is the RIDW result in Raster,
 \code{params} being \code{bestp} is the best distance weighting power,
 \code{MSE} is the Residual Mean squared error of the residuals and
  finally \code{linear_Model} is  the adjusted linear Model.
}
\description{
This function use gstat packages for interpolate spatial point data (\link[sp]{SpatialPoints})
and RasterLayer data (see \link[raster]{raster}).
}
\details{
R_IDW use crossvalidation Leave-p-out cross-validation (LpO CV) and force brute (optimize MSE)
 for estimate the best idp power coeficient.
}
\examples{
library(raster)
library(Dorado)
data("Dorado")
gauge <- mean_doble_Station(gauge = Dorado$gauge,cov = Dorado$TRMM)
sat <- Dorado$TRMM
x <- IDW(gauge = gauge,cov = sat,formula = PP_anual~1)
}
\seealso{
\link[gstat]{idw}
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}
