% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ked.R
\name{KED}
\alias{KED}
\title{KED (Kriging with Extrenal Drift)}
\usage{
KED(gauge, cov, boundariess, formula, fx = c(NA, NA, NA),
  saveparam = "/home/senamhi-cesar/Escritorio/name.Rdata",
  saverast = "/home/senamhi-cesar/Escritorio/name.tif")
}
\arguments{
\item{gauge}{An object of SpatialPointsDataFrame-class.}

\item{cov}{An object of RasterLayer.}

\item{idpR}{Is the range of the power coeficient}
}
\description{
KED (Kriging with Extrenal Drift)
}
\author{
Cesar Aybar
}

