% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandoblestation.R
\name{mean_doble_Station}
\alias{mean_doble_Station}
\title{Mean Doble Station}
\usage{
mean_doble_Station(gauge = NULL, cov = NULL, longlat = TRUE)
}
\arguments{
\item{gauge}{Is an object of SpatialPointsDataFrame-class.}

\item{cov}{Is an object of RasterLayer.}

\item{longlat}{Using WGS84 system}
}
\description{
Mean Doble Station
}
\examples{
library(raster)
library(Dorado)
data("Dorado")
gauge <- mean_doble_Station(gauge = Dorado$gauge,cov = Dorado$TRMM)
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}
