% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_IDW.R
\name{R_IDW}
\alias{R_IDW}
\title{Regression IDW Optimizing inverse distance weighting power}
\usage{
R_IDW(gauge, newdata, formula, idpR = seq(0.8, 3.5, 0.1))
}
\arguments{
\item{gauge}{Is an object of SpatialPointsDataFrame class.}

\item{newdata}{Is An object of RasterLayer.}

\item{formula}{that defines the dependent variable as a linear model
of independent variables; suppose the dependent variable has
name 'z', for ordinary and simple kriging use the formula
'z~1'; for simple kriging also define 'beta' (see below); for
universal kriging, suppose 'z' is linearly dependent on 'x'
and 'y', use the formula 'z~x+y'.}

\item{idpR}{Is vector numeric of the power coeficient to evaluate.}
}
\description{
This function use gstat packages for interpolate spatial point data (see \code{\link[sp]{SpatialPointsDataFrame}} )
and RasterLayer data  (see \code{\link[raster]})
}
\details{
R_IDW use crossvalidation Leave-p-out cross-validation (LpO CV) and force brute (optimize MSE)
 for estimate the best idp power coeficient.
}
\author{
Cesar Aybar
}
\seealso{
\link[gstat]{idw}
}

