% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDW.R
\name{RegressionIDW}
\alias{RegressionIDW}
\title{Regression-IDW optimizado}
\usage{
RegressionIDW(gauge = dperu2, sat = rSAT, formula = gauge ~ cov,
  idpR = seq(0.8, 3.5, 0.1), saveparam, saverast)
}
\arguments{
\item{gauge}{An object of SpatialPointsDataFrame-class.}

\item{idpR}{Is the range of the power coeficient}

\item{cov}{An object of RasterLayer.}
}
\description{
Regression-IDW optimizado
}
\author{
Cesar Aybar
}

