% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleR.R
\name{resampleR}
\alias{resampleR}
\title{Resamplear raster en funcion a otro}
\usage{
resampleR(R1 = NULL, R2 = NULL, t_res = NULL, t_ext = NULL,
  Routput = "~/")
}
\arguments{
\item{R1}{Es un objeto raster que se desea resamplear.}

\item{R2}{Es un objeto raster de la resolucion deseada}

\item{t_res:}{Opcional parametro necesario si no se ingresa raster con que comparar.}

\item{t_ext:}{Opcional parametro necesario si no se ingresa raster con que comparar.}

\item{Routput:}{Es el carpeto de destino del raster}
}
\description{
Resamplear raster en funcion a otro
}
\examples{
#Cuando tienes un raster con que comparar.
resampleR(R1=Prec_Bolivia,R2=dem,Routput = "~/")
#Cuando no tienes un raster con que comparar.
resampleR(R1=Prec_Bolivia,t_res=c(0.1,0.1),t_ext=c(-71,-65,-23,-13),Routput = "~/")
}
\author{
Cesar Luis Aybar Camacho
}

