% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDW.R
\name{RIDW}
\alias{RIDW}
\title{Regression IDW Optimizing inverse distance weighting power}
\usage{
RIDW(gauge, cov, formula, idpR = seq(0.8, 3.5, 0.1), qgis = F)
}
\arguments{
\item{gauge}{Is an object of SpatialPointsDataFrame class.}

\item{formula}{that defines the dependent variable as a linear model
of independent variables; suppose the dependent variable has
name 'z', for Regression Inverse Distance Weigthing (RIDW) use the formula
'z~x+y+....', you do not need define}

\item{idpR}{Is vector numeric of the power coeficient to evaluate.}

\item{newdata}{Is An object of RasterLayer.}
}
\value{
a List that contains: \code{Interpol} is the RIDW result in Raster,
 \code{params} being \code{bestp} is the best distance weighting power,
 \code{MSE} is the Residual Mean squared error of the residuals and
  finally \code{linear_Model} is  the adjusted linear Model.
}
\description{
This function use gstat packages for interpolate spatial point data (see \code{\link[sp]{SpatialPointsDataFrame}} )
and RasterLayer data  (see \code{\link[raster]})
}
\details{
R_IDW use crossvalidation Leave-p-out cross-validation (LpO CV) and force brute (optimize MSE)
 for estimate the best idp power coeficient.
}
\examples{
 library(raster)
 data(Titicaca)
 x <- RIDW(gauge = Titicaca$rain,cov = stack(Titicaca$cov),formula = rain~prec+dem)
 plot(x$Interpol)
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}
\seealso{
\link[gstat]{idw}
}

