% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titicaca.R
\docType{data}
\name{Titicaca}
\alias{Titicaca}
\title{Border Peru Bolivia Rainfall Information
A dataset containing rainfall information and other spatial covariables.
 Titicaca. The variables are as follows:}
\format{A list with 2 camps: rain and cov.
\itemize{
  \item rain: 3 rainfall station is a sp object.
  \item Titicaca$long: longitude raster.
  \item Titicaca$lat: latitude raster.
  \item Titicaca$prec: rainfall raster.
  \item Titicaca$dem: Digital Elevation Model raster.
  \item Titicaca$dist: Distance to sea raster}}
\usage{
Titicaca
}
\description{
Border Peru Bolivia Rainfall Information
A dataset containing rainfall information and other spatial covariables.
 Titicaca. The variables are as follows:
}
\examples{
library(Dorado)
data("Titicaca")
}
\keyword{datasets}

