% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_fit.R
\name{variogram_fit}
\alias{variogram_fit}
\title{Variogram Estimate based in RasterLayer and sp Objects}
\usage{
variogram_fit(gauge, cov, formula, ...)
}
\arguments{
\item{gauge}{Is a object of SpatialPointsDataFrame-class.}

\item{cov}{Is a object of RasterLayer.}

\item{formula}{that defines the dependent variable as a linear model
of independent variables; suppose the dependent variable has
name 'z', for Kriging with Extrenal Drift (KED) use the formula
'z~x+y+....', you do not need define}

\item{\code{...}}{parameters that are passed on to \link[gstat]{variogram} variogram when calculating the sample variogram}
}
\description{
Variogram Estimate based in RasterLayer and sp Objects
}
\examples{
data(Dorado)
variogram_fit(gauge = Dorado$rain,cov = stack(Dorado$cov),rain~prec+dem)
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}
\seealso{
\link[gstat]{fit.variogram} \link[automap]{autofitVariogram}
}

