% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XYraster.R
\name{XY_R}
\alias{XY_R}
\title{Estimate Longitude and Latitude RasterLayer object}
\usage{
XY_R(x)
}
\arguments{
\item{x:}{RasterLayer object}
}
\description{
Estimate Longitude and Latitude RasterLayer object
}
\examples{
data(Prec_Bolivia)
generateXYraster(Prec_Bolivia)
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}

