% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dorado.R
\docType{data}
\name{Dorado}
\alias{Dorado}
\title{Altiplane Peru Bolivia Rainfall Information}
\format{A list with 2 camps: rain and cov:
\itemize{
  \item rain: 72 rainfall station is a sp object.
  \item cov$long: longitude raster.
  \item cov$lat: latitude raster.
  \item cov$prec: rainfall raster.
  \item cov$dem: Digital Elevation Model raster.
  \item cov$dist: Distance to sea raster}}
\usage{
Dorado
}
\description{
A dataset containing rainfall information and other spatial covariables.
 Dorado. The variables are as follows:
}
\examples{
library(Dorado)
data("Dorado")
}
\keyword{datasets}

