% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ked.R
\name{KED}
\alias{KED}
\title{KED (Kriging with Extrenal Drift or Ordinary Kriging) using Raster and sp Objects}
\usage{
KED(gauge, cov, formula, model, crossval = F, ...)
}
\arguments{
\item{gauge}{Is an object of SpatialPointsDataFrame-class.}

\item{cov}{Is an object of RasterLayer.}

\item{formula}{that defines the dependent variable as a linear model}

\item{model}{variogram model of dependent variable (or its residuals),
defined by a call to \code{\link[Dorado]{FitVariogram}}}

\item{crossval}{logical, Generate crosvalidation
of independent variables; suppose the dependent variable has
name 'z', for Kriging with Extrenal Drift (KED) use the formula
'z~x+y+....', you do not need define}

\item{...}{parameters that are passed on to \link[gstat]{variogram} variogram when calculating the sample variogram}
}
\value{
a List that contains: \code{Interpol} is the KED result in Raster,
 \code{params} being \code{residual} that are spatial residual obtains in the cross-validation (residual),
 \code{MSE} is the Residual Mean squared error and finally \code{var} is  the variogram model.
}
\description{
KED (Kriging with Extrenal Drift or Ordinary Kriging) using Raster and sp Objects
}
\examples{
library(raster)
library(Dorado)
data("Dorado")
gauge <- mean_doble_Station(gauge = Dorado$gauge,cov = Dorado$TRMM)
sat <- Dorado$TRMM
x <- KED(gauge = gauge,cov = sat,formula = PP_anual~Precipitacion_Anual,crossval = TRUE)
}
\seealso{
\link[gstat]{krige} \link[automap]{autoKrige}
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}
