% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDW.R
\name{RegressionIDW}
\alias{RegressionIDW}
\title{Regression-IDW optimizado}
\usage{
RegressionIDW(gauge = dperu2, sat = rSAT, idpR = seq(0.8, 3.5, 0.1),
  saveparam, saverast)
}
\arguments{
\item{gauge}{An object of SpatialPointsDataFrame-class.}

\item{sat}{An object of RasterLayer.}

\item{idpR}{Is the range of the power coeficient}
}
\description{
Regression-IDW optimizado
}
\author{
Cesar Aybar
}

