% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_R.R
\name{resampleR}
\alias{resampleR}
\title{Fast Resample RasterLayer object}
\usage{
resampleR(R1 = NULL, R2 = NULL, t_res = NULL, t_ext = NULL,
  Routput = "~/")
}
\arguments{
\item{R1}{Is a RasterLayer object to resample.}

\item{R2}{Is a RasterLayer object that contains the CRS and desired geometry.}

\item{t_res:}{Numeric. Is the pixel resolution (c(xres,yres))}

\item{t_ext:}{Numeric.(c(xmin,ymin,xmax,ymax)). set georeferenced extents of output file to be created (in target SRS).}

\item{Routput:}{Character. The destination file name.}
}
\value{
A RasterLayer object
}
\description{
This function allows manipulating the geometry of one RasterLayer object.
}
\details{
This function use \code{\link[gdalUtils]{gdalwarp}} for resampling RasterLayer object and
implement the two next possible situations. The first is that you have two RasterLayer objects
and you want to change the geometries and SRC of the first using the second. The second
possibility is that you have one RasterLayer object and you know the pixel resolution
and the extents.
}
\examples{
#First Case
resampleR(R1=Prec_Bolivia,R2=dem,Routput = "~/")
#Second Case
resampleR(R1=Prec_Bolivia,t_res=c(0.1,0.1),t_ext=c(-71,-65,-23,-13),Routput = "~/")
}
\author{
Cesar Aybar <aybar1994@gmail.com>
}

